+++ seetm/core/mappers/token_to_token_mapper.py
@@ -30,8 +30,16 @@ class TokenToTokenMapper:
                            "Rule based mappings will be ignored")
             self.token_to_token_map = {}
 
+    def get_mappables(self):
+        return list(self.token_to_token_map.keys()) + list(self.token_to_token_map.values())
+
+    def get_maps(self):
+        return self.token_to_token_map
+
     def map(self, text: Text) -> Text:
         text_copy = deepcopy(text)
+        text_copy = text_copy.strip()
+
         if not self.case_sensitive:
             text_copy = str.lower(text_copy)
 
@@ -40,7 +48,15 @@ class TokenToTokenMapper:
                 token_map = [str.lower(token) for token in token_map]
             token_map = list(set(token_map))
 
-            for pattern in token_map:
-                text_copy = regex.sub(pattern, token_key, text_copy)
+            # sort in decending order to captcher longer patterns first
+            token_map.sort(key=lambda x: len(x), reverse=True)
 
+            for pattern in token_map:
+                if pattern == token_key:
+                    continue
+
+                text_copy = regex.sub('[\s]' + pattern + '[\s]', f" {token_key} ", text_copy)
+                text_copy = regex.sub('^' + pattern + '[\s]', f"{token_key} ", text_copy)
+                text_copy = regex.sub('[\s]' + pattern + '$', f" {token_key}", text_copy)
+                text_copy = regex.sub('^' + pattern + '$', f"{token_key}", text_copy)
         return text_copy